:package: add mut and unmute
